First pass

+++ src/snprintf.c
@@ -339,7 +339,7 @@ static int dopr(char *buffer, size_t maxlen, const char *format, va_list args)
                 break;
             case 'p':
                 strvalue = va_arg(args, void *);
-                total += fmtint(buffer, &currlen, maxlen, (long)strvalue, 16,
+                total += fmtint(buffer, &currlen, maxlen, (size_t)strvalue, 16,
                                 min, max, flags);
                 break;
             case 'n':
